pythonapi/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
mysql-connector-python==8.0.16
apscheduler==3.9.1
requests==2.31.0
# Added for testing and coverage
pytest
pytest-cov
httpx
pytest-asyncio

